+++ src/frontend/indextypes.ts
@@ -1,11 +1,7 @@
 import type { IndexClass } from "@src/frontend/indexing";
-import { GPULinearIndex } from "@src/method/gpulinearindex";
-import { Hybrid } from "@src/method/hybrid";
-import { InvertedIndex } from "@src/method/invertedindex";
 import { LinearIndex } from "@src/method/linearindex";
-import { Ngram } from "@src/method/ngram";
-
-export const HybridBigramInvertedIndex = Hybrid("HybridBigramInvertedIndex", Ngram(2, InvertedIndex), InvertedIndex);
+import { HybridBigramInvertedIndex } from "@src/method/hybridbigraminvertedindex";
+import { GPULinearIndex } from "@src/method/gpulinearindex";
 
 export const IndexTypes: { [key: string]: IndexClass } = {
     LinearIndex,